requirements.txt
asgiref==3.11.1
boto3==1.42.59
botocore==1.42.59
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==40.5.1
jmespath==1.1.0
lorem-text==3.0
mutagen==1.47.0
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
gunicorn==23.0.0
whitenoise==6.9.0
python-dotenv==1.0.1
openpyxl==3.1.5
qrcode[pil]==7.4.2
reportlab==4.0.4
requests==2.31.0
PyPDF2==3.0.1 




this requirement file is  contains all the dependencies for the full project not only for the licenses module so you can use it to install all the dependencies for the full project